+++ Controllers/WydawnictwaController.cs
@@ -2,13 +2,15 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using WebApplication1.Models;
+using System.Security.Claims; // <-- TO JEST POTRZEBNE DO POBRANIA ID UŻYTKOWNIKA
 
 namespace WebApplication1
 {
+    [Authorize]
     public class WydawnictwaController : Controller
     {
         private readonly BibliotekaContext _context;
@@ -18,43 +20,36 @@ namespace WebApplication1
             _context = context;
         }
 
-        // GET: Wydawnictwa
+        [AllowAnonymous]
         public async Task<IActionResult> Index()
         {
             return View(await _context.Wydawnictwa.ToListAsync());
         }
 
-        // GET: Wydawnictwa/Details/5
+        [AllowAnonymous]
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var wydawnictwo = await _context.Wydawnictwa
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (wydawnictwo == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var wydawnictwo = await _context.Wydawnictwa.FirstOrDefaultAsync(m => m.Id == id);
+            if (wydawnictwo == null) return NotFound();
             return View(wydawnictwo);
         }
 
-        // GET: Wydawnictwa/Create
         public IActionResult Create()
         {
             return View();
         }
 
-        // POST: Wydawnictwa/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Nazwa")] Wydawnictwo wydawnictwo)
         {
+            // 1. ZAPISUJEMY KTO TWORZY WPIS
+            wydawnictwo.UserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            
+            // 2. Usuwamy to z walidacji, żeby serwer nie krzyczał, że pole było puste w formularzu
+            ModelState.Remove("UserId");
+
             if (ModelState.IsValid)
             {
                 _context.Add(wydawnictwo);
@@ -64,93 +59,75 @@ namespace WebApplication1
             return View(wydawnictwo);
         }
 
-        // GET: Wydawnictwa/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var wydawnictwo = await _context.Wydawnictwa.FindAsync(id);
-            if (wydawnictwo == null)
+            if (wydawnictwo == null) return NotFound();
+
+            // 3. OCHRONA: CZY TO JEST TWÓJ WPIS?
+            if (wydawnictwo.UserId != User.FindFirstValue(ClaimTypes.NameIdentifier))
             {
-                return NotFound();
+                return Forbid(); // Odcięcie dostępu (Zwraca oficjalny błąd braku uprawnień)
             }
+
             return View(wydawnictwo);
         }
 
-        // POST: Wydawnictwa/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Nazwa")] Wydawnictwo wydawnictwo)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Nazwa,UserId")] Wydawnictwo wydawnictwo)
         {
-            if (id != wydawnictwo.Id)
+            if (id != wydawnictwo.Id) return NotFound();
+
+            // Zabezpieczenie przed atakiem (nie pozwalamy zmienić właściciela)
+            if (wydawnictwo.UserId != User.FindFirstValue(ClaimTypes.NameIdentifier))
             {
-                return NotFound();
+                return Forbid();
             }
+
+            ModelState.Remove("UserId");
 
             if (ModelState.IsValid)
             {
-                try
-                {
-                    _context.Update(wydawnictwo);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!WydawnictwoExists(wydawnictwo.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
+                _context.Update(wydawnictwo);
+                await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
             return View(wydawnictwo);
         }
 
-        // GET: Wydawnictwa/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
-            var wydawnictwo = await _context.Wydawnictwa
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (wydawnictwo == null)
+            var wydawnictwo = await _context.Wydawnictwa.FirstOrDefaultAsync(m => m.Id == id);
+            if (wydawnictwo == null) return NotFound();
+
+            // OCHRONA PRZED USUNIĘCIEM
+            if (wydawnictwo.UserId != User.FindFirstValue(ClaimTypes.NameIdentifier))
             {
-                return NotFound();
+                return Forbid();
             }
 
             return View(wydawnictwo);
         }
 
-        // POST: Wydawnictwa/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var wydawnictwo = await _context.Wydawnictwa.FindAsync(id);
-            if (wydawnictwo != null)
+            
+            // OSTATNIA LINIA OBRONY PRZED USUNIĘCIEM
+            if (wydawnictwo != null && wydawnictwo.UserId == User.FindFirstValue(ClaimTypes.NameIdentifier))
             {
                 _context.Wydawnictwa.Remove(wydawnictwo);
+                await _context.SaveChangesAsync();
             }
-
-            await _context.SaveChangesAsync();
+            
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool WydawnictwoExists(int id)
-        {
-            return _context.Wydawnictwa.Any(e => e.Id == id);
         }
     }
 }
